+++ massa-grpc/src/stream/mod.rs
@@ -2,8 +2,6 @@
 
 /// stream new blocks
 pub mod new_blocks;
-/// unidirectional streaming new blocks
-pub mod new_blocks_server;
 /// stream new endorsements
 pub mod new_endorsements;
 /// stream new blocks headers
@@ -14,8 +12,6 @@ pub mod new_operations;
 pub mod new_slot_abi_call_stacks;
 /// subscribe new slot execution outputs (bi directional streaming)
 pub mod new_slot_execution_outputs;
-/// subscribe new slot execution outputs (unidirectional streaming)
-pub mod new_slot_execution_outputs_server;
 /// subscribe new slot transfers
 pub mod new_slot_transfers;
 /// send_blocks streaming
