+++ massa-grpc/src/stream/new_blocks.rs
@@ -124,3 +124,62 @@ pub(crate) async fn new_blocks(
     // Return the new stream of blocks
     Ok(Box::pin(out_stream) as NewBlocksStreamType)
 }
+
+/// Creates a new stream of new produced and received blocks
+/// uni-directional streaming
+pub(crate) async fn new_blocks_server(
+    grpc: &MassaPublicGrpc,
+    request: Request<grpc_api::NewBlocksRequest>,
+) -> Result<NewBlocksStreamType, GrpcError> {
+    // Create a channel to handle communication with the client
+    let (tx, rx) = tokio::sync::mpsc::channel(grpc.grpc_config.max_channel_size);
+    // Get the inner stream from the request
+    let request = request.into_inner();
+    // Subscribe to the new blocks channel
+    let mut subscriber = grpc.consensus_broadcasts.block_sender.subscribe();
+    // Clone grpc to be able to use it in the spawned task
+    let grpc_config = grpc.grpc_config.clone();
+
+    tokio::spawn(async move {
+        // let filters = match get_filter_new_blocks(request, &grpc_config) {
+        let filter = match FilterNewBlocks::build_from_request(request, &grpc_config) {
+            Ok(filter) => filter,
+            Err(err) => {
+                error!("failed to get filter: {}", err);
+                // Send the error response back to the client
+                if let Err(e) = tx.send(Err(err.into())).await {
+                    error!("failed to send back NewBlocks error response: {}", e);
+                }
+                return;
+            }
+        };
+
+        loop {
+            // Receive a new block from the subscriber
+            match subscriber.recv().await {
+                Ok(massa_block) => {
+                    // Check if the block should be sent
+                    if let Some(data) = filter.filter_output(massa_block, &grpc_config) {
+                        // Send the new block through the channel
+                        if let Err(e) = tx
+                            .send(Ok(grpc_api::NewBlocksResponse {
+                                signed_block: Some(data.into()),
+                            }))
+                            .await
+                        {
+                            error!("failed to send new block : {}", e);
+                            break;
+                        }
+                    }
+                }
+                Err(e) => error!("error on receive new block : {}", e),
+            }
+        }
+    });
+
+    // Create a new stream from the received channel
+    let out_stream = tokio_stream::wrappers::ReceiverStream::new(rx);
+
+    // Return the new stream of blocks
+    Ok(Box::pin(out_stream) as NewBlocksStreamType)
+}
